+++ src/bidding/american.rs
@@ -72,7 +72,7 @@ pub(in crate::bidding) mod notrump;
 pub(in crate::bidding) mod openings;
 mod raises;
 mod rebids;
-mod responses;
+pub(in crate::bidding) mod responses;
 pub(in crate::bidding) mod slam;
 mod strong_two;
 mod weak_twos;